Fix lint configuration
The provider API test modules weren't being recognized as tests by
ESLint. The config has been updated to apply Jest lint rules to them.

A bunch of lint rules have been disabled for now on a per-file basis,
because the structure of the shared tests is incompatible with the
rules. We did something similar on the extension, except less rules
needed to be disabled there. We should give these a pass later to
consider whether we can re-enable any of them.

The only lint fixes made were to comply with the preference for
`toBeUndefined` over `toBe(undefined)`.